+++ include/blosc2.h
@@ -818,7 +818,7 @@ typedef int     (*blosc2_close_cb)(void *stream);
 typedef int64_t (*blosc2_tell_cb)(void *stream);
 typedef int     (*blosc2_seek_cb)(void *stream, int64_t offset, int whence);
 typedef int64_t (*blosc2_write_cb)(const void *ptr, int64_t size, int64_t nitems, void *stream);
-typedef int64_t (*blosc2_read_cb)(void *ptr, size_t size, size_t nitems, void *stream);
+typedef int64_t (*blosc2_read_cb)(void *ptr, int64_t size, int64_t nitems, void *stream);
 typedef int     (*blosc2_truncate_cb)(void *stream, int64_t size);
